Crate zkevm_circuits

zk_evm

Modules

bytecode_circuit

The bytecode circuit implementation.

circuit_tools

Circuit utilities

copy_circuit

The Copy circuit implements constraints and lookups for read-write steps for copied bytes while execution opcodes such as CALLDATACOPY, CODECOPY, LOGS, etc.

evm_circuit

The EVM circuit implementation.

exp_circuit

Exponentiation verification circuit.

instance

The instance definition.

keccak_circuit

The keccak circuit implementation.

mpt_circuit

The MPT circuit implementation.

pi_circuit

Public Input Circuit implementation

root_circuit

The Root circuit implementation.

state_circuit

The state circuit implementation.

super_circuit

The Super Circuit is a circuit that contains all the circuits of the zkEVM in order to achieve two things:

table

Table definitions used cross-circuits

test_util

Testing utilities

tx_circuit

The transaction circuit implementation.

util

Common utility traits and functions.

witness

Witness for all circuits. The Block<F> is the witness struct post-processed from geth traces and used to generate witnesses for circuits.

Macros

_cb

Creates a dummy constraint builder that cannot be used to add constraints.

_ifx

ifx

_matchx

matchx Supports _ which works the same as in the normal match: if none of the other arms are active the _ arm will be executed and so can be used to return some default values or could also be marked as unreachable (using the unreachablex! macro).

_require

_require

_require2

_require2

_unreachablex

Can be used to mark a specific branch as unreachable

assign

assign advice

assignf

assign fixed

circuit

Circuit builder macros Nested macro’s can’t do repetition https://github.com/rust-lang/rust/issues/35853 so we expose a couple of permutations here manually.

concat_with_preamble

Concats arguments with preamble consisting of the originating file and line.

impl_expr

Implementation trait Expr for type able to be casted to u64

impl_expr_result

Implement ExprResult for tupples

impl_expr_vec

Implementation trait ExprVec for type able to be casted to an Expression

matchw

matchw - Resembles matchx so that the witness generation can look like the circuit code.

require_parser

require_parser