Module zkevm_circuits::evm_circuit

The EVM circuit implementation.

Re-exports

pub use self::EvmCircuit as TestEvmCircuit;

pub use crate::witness;

Modules

param

Constants and parameters for the EVM circuit

step

EVM execution state. We model the EVM execution as a finite state machine. The execution of a EVM block goes from one state to another, but never at an undefined state. The EVM circuit enables the selectors and thus activates the constraints for the state that the execution reaches.

table

Fixed lookup tables and dynamic lookup tables for the EVM circuit

Structs

EvmCircuit

Tx Circuit for verifying transaction signatures

EvmCircuitConfig

EvmCircuitConfig implements verification of execution trace of a block.

EvmCircuitConfigArgs

Circuit configuration arguments