Trait zkevm_circuits::table::LookupTable

pub trait LookupTable<F: Field> {
    // Required methods
    fn columns(&self) -> Vec<Column<Any>>;
    fn annotations(&self) -> Vec<String>;

    // Provided methods
    fn advice_columns(&self) -> Vec<Column<Advice>> { ... }
    fn table_exprs(&self, meta: &mut VirtualCells<'_, F>) -> Vec<Expression<F>> { ... }
    fn annotate_columns(&self, cs: &mut ConstraintSystem<F>) { ... }
    fn annotate_columns_in_region(&self, region: &mut Region<'_, F>) { ... }
}

Trait used to define lookup tables

Required Methods

fn columns(&self) -> Vec<Column<Any>>

Returns the list of ALL the table columns following the table order.

fn annotations(&self) -> Vec<String>

Returns the String annotations associated to each column of the table.

Provided Methods

fn advice_columns(&self) -> Vec<Column<Advice>>

Returns the list of ALL the table advice columns following the table order.

fn table_exprs(&self, meta: &mut VirtualCells<'_, F>) -> Vec<Expression<F>>

Return the list of expressions used to define the lookup table.

fn annotate_columns(&self, cs: &mut ConstraintSystem<F>)

Annotates a lookup table by passing annotations for each of it’s columns.

fn annotate_columns_in_region(&self, region: &mut Region<'_, F>)

Annotates columns of a table embedded within a circuit region.