Module zkevm_circuits::util

Common utility traits and functions.

Modules

cell_manager

Cell Manager

cell_placement_strategy

Cell Placement strategies

int_decomposition

Define IntDecomposition to decompose int into byte limbs

word

Define generic Word type with utility functions

Structs

Challenges

All challenges used in SuperCircuit.

U256

Little-endian large integer type 256-bit unsigned integer.

Traits

Expr

Trait that implements functionality to get a constant expression from commonly used types.

SubCircuit

SubCircuit is a circuit that performs the verification of a specific part of the full Ethereum block verification. The SubCircuit’s interact with each other via lookup tables and/or shared public inputs. This type must contain all the inputs required to synthesize this circuit (and the contained table(s) if any).

SubCircuitConfig

SubCircuit configuration

Functions

log2_ceil

Ceiling of log_2(n)

query_expression

Steal the expression from gate

Type Definitions

Address