Module zkevm_circuits::witness

Witness for all circuits. The Block<F> is the witness struct post-processed from geth traces and used to generate witnesses for circuits.

Structs

Block

Block is the struct used by all circuits, which contains all the needed data for witness generation.

BlockContext

Block context for execution

Call

Circuit Input related to an Ethereum Call

ExecStep

An execution step of the EVM.

MptUpdate

An MPT update whose validity is proved by the MptCircuit

MptUpdateRow

The field element encoding of an MPT update, which is used by the MptTable

MptUpdates

All the MPT updates in the MptCircuit, accessible by their key

RwMap

Rw constainer for a witness block

RwRow

Rw table row assignment

Transaction

Result of the parsing of an Ethereum Transaction.

Enums

Rw

Read-write records in execution. Rws are used for connecting evm circuit and state circuits.

Functions

block_convert

Convert a block struct in bus-mapping to a witness block used in circuits