Struct zkevm_circuits::witness::ExecStep

pub struct ExecStep {

    pub exec_state: ExecState,
    pub pc: u64,
    pub stack_size: usize,
    pub memory_size: usize,
    pub gas_left: u64,
    pub gas_cost: u64,
    pub gas_refund: u64,
    pub call_index: usize,
    pub rwc: RWCounter,
    pub reversible_write_counter: usize,
    pub reversible_write_counter_delta: usize,
    pub log_id: usize,
    pub bus_mapping_instance: Vec<OperationRef, Global>,
    pub copy_rw_counter_delta: u64,
    pub error: Option<ExecError>,

}

An execution step of the EVM.

Fields
exec_state: ExecState

Execution state

pc: u64

Program Counter

stack_size: usize

Stack size

memory_size: usize

Memory size

gas_left: u64

Gas left

gas_cost: u64

Gas cost of the step. If the error is OutOfGas caused by a “gas uint64 overflow”, this value will not be the actual Gas cost of the step.

gas_refund: u64

Accumulated gas refund

call_index: usize

Call index within the Transaction.

rwc: RWCounter

The global counter when this step was executed.

reversible_write_counter: usize

Reversible Write Counter. Counter of write operations in the call that will need to be undone in case of a revert. Value at the beginning of the step.

reversible_write_counter_delta: usize

Number of reversible write operations done by this step.

log_id: usize

Log index when this step was executed.

bus_mapping_instance: Vec<OperationRef, Global>

The list of references to Operations in the container

copy_rw_counter_delta: u64

Number of rw operations performed via a copy event in this step.

error: Option<ExecError>

Error generated by this step

Implementations

impl ExecStep

pub fn new( step: &GethExecStep, call_ctx: &CallContext, rwc: RWCounter, reversible_write_counter: usize, log_id: usize ) -> ExecStep

Create a new Self from a GethExecStep.

pub fn oog_or_stack_error(&self) -> bool

Returns true if error is oog and stack related..

pub fn opcode(&self) -> Option<OpcodeId>

Try get opcode, if possible

pub fn rw_index(&self, index: usize) -> OperationRef

get rw index

pub fn rw_indices_len(&self) -> usize

Get the size of read and writes

pub fn stack_pointer(&self) -> u64

Get stack pointer

pub fn memory_word_size(&self) -> u64

The memory size in word before this step